Arreglando para pruebas con postman

+++ boda-back/src/main/java/co/edu/uniandes/csw/boda/entities/TarjetaCreditoEntity.java
@@ -7,10 +7,14 @@ import javax.persistence.Temporal;
 import javax.persistence.TemporalType;
 import uk.co.jemos.podam.common.PodamExclude;
 
+/**
+ *
+ * @author ca.guerrero
+ */
 @Entity
 public class TarjetaCreditoEntity extends BaseEntity {
     private Long numero;
-    private Double numDeSeg;
+    private int numDeSeg;
     @Temporal(TemporalType.DATE)
     private Date fechaVen;
     
@@ -42,11 +46,11 @@ public class TarjetaCreditoEntity extends BaseEntity {
         this.numero = numero;
     }
 
-    public Double getNumDeSeg() {
+    public int getNumDeSeg() {
         return numDeSeg;
     }
 
-    public void setNumDeSeg(Double numDeSeg) {
+    public void setNumDeSeg(int numDeSeg) {
         this.numDeSeg = numDeSeg;
     }
 }
